V5:   - PrimitiveColumn: Make Where(Span) throw when not implemented yet.   - Make IndexSet take int length, not uint (annoying to cast everywhere).   - Fix WhereSingle bug which didn't check the last partial page.   - Add IndexSet_Where tests for multiple types, OR, AND NOT

+++ V5/V5.ConsoleTest/Program.cs
@@ -30,7 +30,7 @@ namespace V5.ConsoleTest
             this.ResponseBytes = new PrimitiveColumn<int>("ResponseBytes", new int[capacity]);
         }
 
-        public uint Count => (uint)this.EventTime.Count;
+        public int Count => this.EventTime.Count;
 
         public void Index(Random r)
         {
@@ -135,7 +135,7 @@ namespace V5.ConsoleTest
         static void PerformanceTests()
         {
             int iterations = 100;
-            uint size = 8 * 1000 * 1000;
+            int size = 8 * 1000 * 1000;
 
             int sum;
             IndexSet set = new IndexSet(size);
